python_code/requirements.txt
pymongo
logging
logmatic-python